Update patch set 3
Patch Set 3:

> Patch Set 3: Verified+1 Code-Review-1
> 
> As suggested by https://swiftshader-review.googlesource.com/c/SwiftShader/+/31612/6/src/Vulkan/VkObject.hpp#32 Thanks Antonio for the pointers!
> 
> I'm trying to wrap my head around the actual practical benefit that could be achieved from this for this exact case. Perfect forwarding is clearly very useful to prevent near-duplicate code when you want/need to accept various forms of parameter passing without extra overhead. In this case though there's exactly one vkCreate* for each Vulkan object type (and even if there were multiple ones they're likely to still have exactly the same parameters as the object's constructor).
> 
> In all this feels like a premature micro-optimization to me, with a small but perceptible negative impact on readability (the std::forward<> is kind of ugly when all you really want to do is write the argument).
> 
> So I'm inclined not to land this. Like so many things with C++, it can be really really useful in specific situations (like in this case the implementation of make_unique<> or other template libraries where you have no control over all its uses and every cycle counts), but this doesn't appear to be such a case. Feel free to prove me wrong.

I understand that this may seem ugly, but it's idiomatic of modern C++ code that forwards arguments, especially variadic arguments. What seems ugly at first becomes perfectly normal as you encounter it more and more often. The advantage is that if ever we want to pass through an r-value that's expensive to copy, by making sure to use perfect forwarding, we will not incur any copies in the "in between" forwarding functions. For example, you'd be able to pass in a std::vector of values through, if you like, and if you pass it with std::move() to the first function in the chain, it will be automatically moved all the way through till it's stored, with no extra copy.

The truth is, not putting the r-value references and std::forward looks like a bug. Normally, if you really want to force a copy, you don't put the && and std::forward and add a comment to explain why you'd want to do that (there are some legitimate use-cases).

Patch-set: 3
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>