Fix alignment of VkWrapper
The alignment of VkWrapper is currently set to match the alignment of
uint64_t which isn't always correct. On some platforms (e.g. Linux x86)
the alignment of uint64_t inside and outside of a struct differ (8 vs
4). Using an alignment of 8 will cause swiftshader to miscalculate the
size of structs containing VkWrappers and the offset of different
fields into the struct. This is why the alignment of VkWrapper should
match that of uint64_t inside a struct, not outside of it.

Bug: chromium:967831
Change-Id: Ibf8240d13406413481cf87e9d45df4478c35be2f