Fix initializing bits after bitfields
SpirvShader::InterfaceComponent is used as part of the state that
comprises the lookup key for cached Reactor routines. It uses 1-bit
bitfields for Boolean values, and the bits following it for padding may
be left uninitialized by the compiler. Computing the hash of all the
state is done by XORing together the memory, so it shouldn't have any-
thing uninitialized. While we memset all of SetupProcessor::State to 0,
the InterfaceComponent state copied in still contained the uninitialized
bits. Computing a hash from it was detected by MemorySanitizer.

Bug chromium:967824

Change-Id: If1676831fc838bf2a6ca869d13e6f6d2052fa6d8