Update patch set 9
Patch Set 9: Code-Review-1

Sorry for the late feedback on this. It's a highly useful experiment that needed to be tested out. But there are other approaches to solving this problem that also should be looked into before we decide on a longer term solution. My worry is that gather is regarded as a silver bullet, when it's really almost no faster than scalar per-component extract/load/insert. Also, the vast majority of texture sampling operations use implicit LOD, so we should avoid regressing performance of that. While that's not a goal of Pastel, it will be soon and there are less intrusive workarounds for the current bugs.

Patch-set: 9
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1