[COPY] Cache callstack results to avoid expensive lookup
This is a forked rebase of change:
https://swiftshader-review.googlesource.com/c/SwiftShader/+/32369

This works well as our use-case is to lookup the same addresses multiple times (loops, helper functions, etc.).
This optimization is especially important for Mac where resolving stack addresses takes much longer than on other
platforms. With the optimization, running a specific gles-unittests test, it was able to spit out the callstack
trace in about 30 seconds; without, it was still emitting jit code after 10 minutes at which point I gave up.

Change-Id: I3eb9b7772f584382d564240bfda33a6940e42086