Don't let shaders negate samplers
Prior to this change, the OpenGL compiler accepted all "non-struct"
"non-array" types for unary operations (++, --, and -). This ignored the
existence of sampler objects. If someone tried to negate a sampler, then
the compiler would tell the assembler to negate a non-numerical type.

This change makes it so the GSL compiler only accepts unary operations
on numerical types (including vectors and matrices).

Also added a unittest that makes sure we gracefully fail to compile
bad unary operations.

Bug chromium:910883