ComputeProgram: Remove data field to fix race.
data is the 0th argument to the coroutine.
There's no good reason why this was sorted as a field when all other arguments were fetched in ComputeProgram::emit().

Having this as a field caused a data race in the rr::Variable materializing, as the destructor of this field was called outside the reactor mutex.

Test: dEQP-VK.synchronization.internally_synchronized_objects.pipeline_cache_compute
Bug: b/133127573
Change-Id: I44545b71714fdebd8f1150bf7d6325f7d4d4d3eb