Prevent integer-overflow on scissor test
Currently we don't restrict any of the parameters (x, y, width, height)
to glScissor aside from forcing width and height to be non-negative.
This means large parameters can cause overflows during the scissor test.

Fortunately, we can avoid overflows when the parameters are set because
there is a maximum viewport size. If x or y are larger than the max
viewport size, then we already don't render anything. So we set
width/height to 0. Otherwise, we clamp height and width are larger to
the maximum viewport size.

Bug chromium:969071

Change-Id: Icb6992551aed6e3e086f96d89931eae0e99899a4