Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> 1) Any code like this should be in a validation layer, or, if that's really not possible, debug only. Note that we are unlikely to merge code like this in master as it adds unnecessary complexity. You may want to think about adding and maintaining a separate branch for validation code, if you really want to merge it.

I tried validation layer with swiftshader. it doesn't report any errors. Not sure if it is a swiftshader bug, or validation layer doesn't validate this kind of problems.

For nvidia vulkan implementation, validation layer can not enabled, because xlib surface extension is not supported with validation layer.

> 2) Why not just use a std::shared_ptr?
ScopedRef is just for increasing and decreasing ref for a vk object in a scope, it is just for debugging. the object can still be deleted with a non-zero ref.

Patch-set: 2
Reviewer: Gerrit User 21844 <21844@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>