Move libvk_swiftshader.lds from sources to inputs
Needed to fix this build error with ToT gn:
https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket.appspot.com/8911014495686240880/+/steps/analyze/0/stdout

Gn does not know how to add a "*.lds" file in sources, and adding
"libvk_swiftshader.lds" to sources does not have any effect on the generated
ninja files.  With ToT gn, this becomes an error.  This CL changes the file to
be added to inputs, which will cause the target to get relinked when the file
changes.

BUG=chromium:964411
R=capn

Change-Id: Ib931f3ec886f06ad53cdc3b89d78e973e6b0aae3