Fix 'control reaches end of non-void function' warnings treated as errors
All switch control flows return, but the compiler doesn't seem smart enough to notice this.

This magically started happening when I switched to using ninja, possibly suggesting that it is either suddenly using a different compiler or the flags are different.

While something I'll try investigating, here's a fix all the same.

Bug: b/130335507
Change-Id: I3c6d00c00ecd03bbd34d51ac62d02fefdc026526