Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)
> 
> > Patch Set 6:
> > 
> > (1 comment)
> > 
> > What ensures that a returned routine survives long enough to be used?
> 
> Nothing. If over 1024 different routines are required between the code generation and the execution of the code, then it would probably crash. A shader should get invalidated if one of its sampler routines get evicted, but it isn't the case currently. I can look into how these things can be sync'd together.

Note: I could put the SpirvShader object pointer into a constant and pass it as an argument to getImageSampler, then using it to bind the routine and unbind all the used routines when the SpirvShader object gets deleted. It might add some cost to the call, though, so I'll put this on hold and think about it a little more.

Patch-set: 6
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>