Parse 'Sample' image instruction operand
OpImageFetch/Read/Write can have an optional 'Sample' operand which
indicates the sample index to operate on. It is orthogonal to other
operands so it is indicated by a new Boolean field in
sw::SamplerFunction and SpirvShader::ImageInstruction. The offset
operand was also turned into a Boolean field / component count.

The sample index is not wired up to be taken into account by the Fetch
sampling routine yet. OpImageRead already had support for the Sample
operand, but OpImageWrite does not.

Bug: b/135265531
Tests: dEQP-VK.pipeline.multisample.sampled_image.*
Change-Id: I20f50a888436775b996221e8283a4c4ab7f28e17