Update patch set 3
Patch Set 3: -Code-Review

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > I'm afraid I think this is really a bad direction to go in.
> > You're creating a helper base class that obfuscates the fact that you're memset()ing a derived class (along with any other base classes of the derived type), all for the sake of broken initialization.
> > 
> > memset(this) is a bunch of bugs waiting to happen. While you're proposed TODO for is_standard_layout should avoid issues like corrupting v-tables, it doesn't avoid corruption of non-POD types. Note, this will not fire any assert:
> > 
> >     #include <type_traits>
> >     #include <string>
> >     #include <vector>
> > 
> >     class C {
> >         std::string s;
> >         std::vector<int> v;
> >     };
> > 
> >     static_assert(std::is_standard_layout<C>::value, "Oh noes");
> 
> For what it's worth, it does fire on MSVC. But even if it didn't, it seems safe to me to zero the underlying memory before their constructor is run?
> 
> > This isn't just a theoretical issue. See: https://devblogs.microsoft.com/oldnewthing/?p=95397
> 
> I believe in that example they're zeroing the string object after construction?

Right - I was thinking of cases where the CacheKey is not the first constructor, but I see you have a regular assert for that. If we go with this, perhaps we can make that a static_assert?

If we are sure there's no way for any of the derived type's fields to be splatted after initialization, and we can also assert the type doesn't have vtables, I'll be happy.
I'd still like to better understand the actual initialization issue this is all trying to fix.

Patch-set: 3
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=0