Update patch set 3
Patch Set 3:

> > I believe in that example they're zeroing the string object after construction?
> 
> Right - I was thinking of cases where the CacheKey is not the first constructor, but I see you have a regular assert for that. If we go with this, perhaps we can make that a static_assert?

I think that requires a dummy non-null constant pointer, but yes it might be doable with static_assert.

> If we are sure there's no way for any of the derived type's fields to be splatted after initialization, and we can also assert the type doesn't have vtables, I'll be happy.

Cool. Unfortunately it looks like CacheKey can be added as a second base class without triggering asserts. That's because it's zero-sized.

I'll try to write a static_assert that if it *were* to have non-zero size that it's at the base address.

> So... this is all based around the use of memcmp in State::operator ==?

No, to quickly compare the keys we compute hashes of them in some cases (e.g. PixelProcessor::States::computeHash()). That treats it like flat memory, so we can't have uninitialized bits anywhere.

Of course we could instead compute hashes based just on the fields, but that's going to be slower and it's easy to forget updating the hash calculation.

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>