Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)
> 
> I'm not sure this has any effect. The SpirvRoutine is destroyed right after emitEpilog (at least for ComputeProgram), along with these variables. If for graphics routines it leads to needless materializations we should check if we can limit the SpirvRoutine object's lifetime there as well.

> I'm not sure this has any effect

I wouldn't put this up for review if I didn't think it did anything. :)

This affects vertex shaders, and possibly fragment shaders too. Due to the way I structured ComputeProgram, you're right in that it probably wouldn't do anything there.

For our old friend all_shared_buffer.48, this change removes around 40k pointless LLVM IR instructions, consisting of allocas and stores that are never read.

The performance gains for LLVMReactor are minor as LLVM's optimizer is smart enough to drop the stores and unused allocas, and so was not worth boasting about in the CL description - that said I think this is worth landing as it means the IR is far less noisy, and I can actually track what's going on. Also assuming we resurrect the Subzero backend, these stores won't be automatically optimised away, and would have a real impact to performance.

> we should check if we can limit the SpirvRoutine object's lifetime there as well.

I did try this, but I stopped going down this path after I realised the repercussions. Let me walk you though the problem:

1. The VertexRoutine constructor calls SpirvShader::emitProlog().
2. The VertexProgram constructor initializes the SpirvRoutine with input builtins.
3. VertexRoutine::generate() sets up a reactor do-while loop for processing the vertices, and an if-block that contains the call program(batch). This is implemented in VertexProgram::program(), and calls SpirvShader::emit() and SpirvShader::emitEpilog().
4. After the call to program(batch), there's the call to computeClipFlags(), and writeCache(), and the if-block scope is closed, this forcefully materializes all variables.

So, we have a bunch of setup and initialization (1, 2) way before the VertexRoutine::generate() loop (3). We have a whole lot of temporary variables generated in SpirvShader::emit() called by program(batch) (3), which are automatically materialized as soon as the if-statement block is closed (4).

I tried moving the SpirvRoutine lifetime scope to inside the if-block, but this immediately became obviously impractical as as we'd end up re-initializing all the builtins & input variables with each iteration of the loop.

I think we may want to revisit the responsibilities and data relationships between Routines/Programs, SpirvShader and SpirvRoutine, but this change seemed like the smallest delta to fix some real nasty codegen.

Patch-set: 2
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>