Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks Paul. What project is affected by this? Do you have a (public) tracker bug for it?

I plan to build SwiftShader for Windows, Mac, and Linux, zip up the builds (the directory containing the libs plus the JSON file), and then download and extract the zips and use the ICD as part of the GraphicsFuzz project.

https://github.com/google/graphicsfuzz

I can always create the JSON manifest manually, but I figured it would be good to make the change in the upstream project if possible. However, it sounds like it might not work reliably on all platforms. Unless the issue was caused by using "./" on Windows, in which case maybe this change will work correctly. My interpretation of the Vulkan Loader docs was that, in order to be interpreted as a relative path, the path must include the correct directory separator and must not start with the directory separator, but I have only tested it on my Linux machine.

Patch-set: 1
Reviewer: Gerrit User 7165 <7165@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>