NFR: Test new SpirvShader inliner - disable the spirv-opt inliner
Inlining doesn't always happen (see b/133213304), which is a requirement for SpirvShader.
While a fix could be made to spirv-opt, I think we can get funtion inlining implemented with minor changes to SpirvShader.

This change is for regres so that I can see which dEQP tests make function calls, and monitor progress with a new implementation.

Bug: b/133213304
Change-Id: I90cbd94020027f65b5f702cdcb25f475c5b1b268