Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Thanks Paul. What project is affected by this? Do you have a (public) tracker bug for it?
> 
> I plan to build SwiftShader for Windows, Mac, and Linux, zip up the builds (the directory containing the libs plus the JSON file), and then download and extract the zips and use the ICD as part of the GraphicsFuzz project.

Thanks for the clarification. Is there anything we can do to serve your needs better for a proper build integration?

We used to have prebuilt version of SwiftShader libraries in the past, but what ended up happening is that bugs got filed by people using outdated versions. Often the bug was already fixed in later revisions, but it takes valuable time to check that.

Also, with a tighter integration it should be easier to pinpoint regressions.

Patch-set: 2