Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Thanks Paul. What project is affected by this? Do you have a (public) tracker bug for it?
> > 
> > I plan to build SwiftShader for Windows, Mac, and Linux, zip up the builds (the directory containing the libs plus the JSON file), and then download and extract the zips and use the ICD as part of the GraphicsFuzz project.
> 
> Thanks for the clarification. Is there anything we can do to serve your needs better for a proper build integration?
> 
> We used to have prebuilt version of SwiftShader libraries in the past, but what ended up happening is that bugs got filed by people using outdated versions. Often the bug was already fixed in later revisions, but it takes valuable time to check that.
> 
> Also, with a tighter integration it should be easier to pinpoint regressions.

Thanks for asking. There are two things I can think of. 

If it is possible to store the binaries for Linux, Mac, and Windows, for every commit to master (or certain commits), such that they are accessible externally (but not necessarily advertised externally), that would potentially be very useful! Of course, there are often challenges in making sure we have a compatible configuration. E.g. which compiler, Debug/Release, etc., especially when dealing with shared libraries.

The other thing that would probably be helpful is to add more configurations to the presubmit tests (but I know this can be a very time-consuming process) or to simply ensure that we use the same configurations as you. E.g. the version of Ubuntu, version of Visual Studio, version of macOS/XCode. It may be that our configurations are already similar to yours, but I know at least that we build with Visual Studio 2015 and you use 2017. You could say you only support, say, Visual Studio 2017, and therefore we should build using that (we are flexible). I think we will mainly be using SwiftShader with Amber.

https://github.com/google/amber

So we ideally just need to build both projects with the same build configurations.

Patch-set: 2
Reviewer: Gerrit User 7165 <7165@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>