PipelineCache implementation
The PipelineCache class now contains members for caching programs and
shaders, which allows significant speedups when using it to prevent
recompiling the same shaders over and over again.

For now, each PipelineCache object contains its own cache and the
cache is not shared between PipelineCache objects. It remains to be
seen if SwiftShader would benefit from always caching shaders and
programs in a global cache.

Notes:
- Merging 2 PipelineCache objects was also implemented
- Added a few "const" where appropriate
- Removed descriptorSets members from ComputeProgram and EmitState,
  as they were unused.

IMPORTANT NOTE: This cl DOES NOT allow pipeline caches to be saved
                and loaded through the pipeline cache data.

Bug b/123588002

Change-Id: I95b183033c03e114d69d4432e5831e26be477033