Reactor: Rework optimization flags.
Remove the old 'runOptimizations' flag on Nucleus::acquireRoutine() and Nucleus::acquireCoroutine().
runOptimizations was always true, and there is already the rr::optimization extern to control the optimization passes performed.

Add a new OptimizationLevel enumerator, and add it to the Function and Coroutine constructors. This is passed down to the Nucleus constructor, which is used to control the optimization level for the JIT.

Unfortunately the LLVMReactorJIT is a lazily constructed singleton, which contains the llvm::orc::ExecutionSession. This needs to be passed the optimization level on construction, so currently the optimization level is fixed to the value passed to the first call of the Nucleus constructor.

All Function and Coroutines are now passed the new vk::ReactorOptimizationLevel. This will be changed to a non-Default value in another CL.

Bug: b/135609394
Change-Id: Ia3e832ccb7f38e980bf5b9caffdc6f7dea072bfe