Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+2
> 
> Ah a classic, that has stung me many a time (the last time just 8 months ago: https://github.com/google/gapid/commit/acb07b377da80bd97049ba9eb4c4ded29444f9b1).
> Sorry for not noticing this earlier.
> 
> Note that llvm::Module::setDataLayout() is usually paired with a call to llvm::Module::setTargetTriple() - however the LLVM Kaleidoscope sample only appears to call the former. Either way, we may want to consider adding:
> 
>   ::module->setTargetTriple(targetMachine->getTargetTriple().str());
> 
> below this new line.

I think the target triple is more of a high-level stand-alone compiler thing. It's used to imply things like -march and -attrs and the object file format to use, but each of these can be controlled more precisely with the ExecutionEngine methods.

Anecdotally note that we make Subzero emit ELF binary format, on every platform. I wrote a minimal ELF loader for it, that works fine on Windows too.

Anyway, I think the target triple that LLVM ends up using is controlled by the config headers that we generate. Seems sane to me. That said, b/136095012 appears to have some weird platform dependence...

Patch-set: 3