Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> If we think of Function<> and Coroutine<> as implementation-independent blueprints of the code to generate (cf. source files), then I don't think we should specify the optimization level for them at constructor time. 

Yup, I had the same thoughts, but unfortunately the way LLVMReactor is currently structured, this cannot currently be pushed any later than the constructor of Nucleus:
The Nucleus constructor builds the LLVMReactorJIT, which in turn builds the llvm::TargetMachine. The target machine holds the JIT optimisation level. The target machine is has an platform dependent implementation, and so it cannot be assumed that we can successfully change the optimization level after it has been constructed.
Unless we do moderately substantial reworkings of LLVMReactor.cpp (which I think we should do, but not now), this is a pretty hard limitation.

> Also I think it can be confusing to have both an argument list corresponding to that for the routine, and arguments that are passed to the back-end.

A confusion that stems from the use of inheritance from rr::Function, which I'm not entirely convinced is a great pattern. Anyway, I'm not opposed to moving this to a setOptimizationLevel() method of rr::Function & rr::Coroutine.

> Just like with static compilation, specifying the optimization level globally seems more convenient to me. 

I'd really prefer to steer away from more global state. It raises questions of initialization order, whether it is legal to change it later, whether it requires thread synchronization, etc etc.
Plus, this approach prevents any sort of per-compile optimization flags, required for VK_PIPELINE_CREATE_DISABLE_OPTIMIZATION_BIT and any background optimizing of the hottest shaders.

> It also ensures that we only have to turn one dial to affect all JIT compilation. 

The dial is rr::ReactorOptimizationLevel. Turned in the next change.

> Overriding that global setting at acquireRoutine() time also seems reasonable to me since we're more explicitly talking to the JIT there.

The override is the rr::Function constructor with the default parameter set to  rr::OptimizationLevel::Default. As I said, I also think that specifying the optimization level at acquireRoutine() is preferable, but requires a refactoring of the way LLVMReactor is structured.

If you would be happy with me doing the restructuring of LLVMReactor, then I will do this to move the opt flag to acquireRoutine(). I would prefer to do this than pollute Reactor with more dodgy global state.

Patch-set: 2
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>