WIP: Refactor LLVMReactor
Split the scattered LLVM global variables into 3 new classes:
• JITGlobals is a singleton that holds all the immutable machine specific
  information for the host device.
• JITBuilder holds all the LLVM state for building routines.
• JITSession holds a LLVM JIT session, compiler and object layer.
  A JITSession is created for each acquired rr::LLVMRoutine as each routine
  may want different target machine settings, and no Reactor routine
  directly link against another. The JITSession is reference counted, and is
  automatically deleted once the last LLVMRoutine is released.

This reduce all the global varibles to down to one - 'jit' a JITBuilder.
In theory, this can be marked as TLS to remove the global reactor mutex lock.

These change also make it possible to specify per-function JIT optimization
settings (not implemented in this change).

Change-Id: If4c90de444830f398eb79081fd349f313a8d3bd7