Reactor: Rework optimization flags.
Replace the old 'runOptimizations' flag on Nucleus::acquireRoutine() and Nucleus::acquireCoroutine() with a new OptimizationLevel enumerator.
runOptimizations was always true, and there is already the rr::optimization extern to control the optimization passes performed.

All Function and Coroutines are now passed the new vk::ReactorOptimizationLevel. This will be changed to a non-Default value in another CL.

Bug: b/135609394
Change-Id: I1154da05d413b18a471a3818fbb03f356a3d0e6c