Create worker threads at device creation time
This prevents costly thread initialization and memory allocations from
happening at draw time.

Deferred worker thread creation was introduced in 7381c99 as part of
crbug.com/336438 to avoid some regression in Chrome's startup time. An
OpenGL context has to be created just for querying some strings and
capabilities. This isn't an issue with Vulkan, so we can eagerly create
the worker threads.

Bug: b/137093675
Change-Id: I918df86f69c5ead887bdcbca4f0d5bfc66da3ac8