Update patch set 1
Patch Set 1:

(2 comments)

Zero objection to landing this as-is. The code works as intended as far as I can tell, and we're too close to release to make iterations on it.

Chris to +2 if it fixes the memory bloat issue.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>