Reactor: Use std::shared_ptr to ref-count rr::Routines
The intrusive reference counting had no smart pointer, making it difficult to correctly manage reference counts. Most of the codebase did not call bind() and unbind() explictly, and relied on some other mechanism for keeping the routine alive.

By switching to std::shared_ptr, we automatically get guarantees that routine cache eviction will not destroy the routine while it is in use.

Bug: b/137524292
Change-Id: I38b5e8ba3ee084572a427a1de20f4f017ceaae5a