vk::Queue: Lazily construct Renderers
They're big and take a while to initialize. Only construct them if we need them.
This also pushes the construction onto the queue, speeding up the main thread.

Finally, this means that all the calls into Renderer come from the same thread
(the vk::Queue's), which is handy for ensuring a yarn::Scheduler is bound.

Bug: b/139142453
Change-Id: I2d1d25f24162339de62cea4d946532b1a2e346ca